+++ src/core/app.module.core.ts
@@ -3,6 +3,7 @@ import { ConfigModule } from '@nestjs/config';
 import { PassportModule } from '@nestjs/passport';
 import { ServeStaticModule } from '@nestjs/serve-static';
 import { TerminusModule } from '@nestjs/terminus';
+import { join } from 'path';
 
 import { AuthController } from '../api/auth.controller';
 import { ChatsController } from '../api/chats.controller';
@@ -21,6 +22,8 @@ import { VersionController } from '../api/version.controller';
 import { WhatsappConfigService } from '../config.service';
 import { SessionManager } from './abc/manager.abc';
 import { WAHAHealthCheckService } from './abc/WAHAHealthCheckService';
+import { DashboardConfigServiceCore } from './config/DashboardConfigServiceCore';
+import { SwaggerConfigServiceCore } from './config/SwaggerConfigServiceCore';
 import { WAHAHealthCheckServiceCore } from './health/WAHAHealthCheckServiceCore';
 import { SessionManagerCore } from './manager.core';
 
@@ -30,15 +33,26 @@ export const IMPORTS = [
   }),
   ServeStaticModule.forRootAsync({
     imports: [],
-    extraProviders: [WhatsappConfigService],
-    inject: [WhatsappConfigService],
-    useFactory: (config: WhatsappConfigService) => {
-      return [
+    extraProviders: [WhatsappConfigService, DashboardConfigServiceCore],
+    inject: [WhatsappConfigService, DashboardConfigServiceCore],
+    useFactory: (
+      config: WhatsappConfigService,
+      dashboardConfig: DashboardConfigServiceCore,
+    ) => {
+      const options = [
+        // Serve files (media)
         {
           rootPath: config.filesFolder,
-          serveRoot: config.files_uri,
+          serveRoot: config.filesUri,
         },
       ];
+      if (dashboardConfig.enabled) {
+        options.push({
+          rootPath: join(__dirname, '..', 'dashboard'),
+          serveRoot: dashboardConfig.dashboardUri,
+        });
+      }
+      return options;
     },
   }),
   PassportModule,
@@ -67,6 +81,8 @@ const PROVIDERS = [
     provide: WAHAHealthCheckService,
     useClass: WAHAHealthCheckServiceCore,
   },
+  DashboardConfigServiceCore,
+  SwaggerConfigServiceCore,
   WhatsappConfigService,
   ConsoleLogger,
 ];
@@ -76,4 +92,6 @@ const PROVIDERS = [
   controllers: CONTROLLERS,
   providers: PROVIDERS,
 })
-export class AppModuleCore {}
+export class AppModuleCore {
+  constructor(protected config: WhatsappConfigService) {}
+}