+++ src/core/engines/webjs/session.webjs.core.ts
@@ -74,17 +74,19 @@ export class WhatsappSessionWebJSCore extends WhatsappSession {
   }
 
   protected getClientOptions(): ClientOptions {
+    // Folder with current file
+    const path = __dirname;
     return {
       puppeteer: {
         headless: true,
         executablePath: this.getBrowserExecutablePath(),
         args: this.getBrowserArgsForPuppeteer(),
       },
-      webVersion: '2.2411.2',
+      webVersion: '2.2412.54',
       webVersionCache: {
-        type: 'remote',
-        remotePath:
-          'https://raw.githubusercontent.com/wppconnect-team/wa-version/main/html/{version}.html',
+        type: 'local',
+        path: path,
+        strict: true,
       },
     };
   }